requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
flasgger==0.9.7.1
pymongo==4.6.0
flask-cors==5.0.0
gunicorn==21.2.0

